other updats for pypi

+++ pypi.sh
@@ -2,8 +2,17 @@
 
 DIST=dist
 PACKAGE=cheap_pie
+PIP=pip3
 
-rm -rf $DIST # remove old package
-python3 setup.py sdist # create package
-twine upload ${DIST}/* # upload to testpypi
-python3 -m pip install ${PACKAGE} # install
+ARGV=$1
+if [ -z "$ARGV" ]
+then
+    rm -rf $DIST # remove old package
+    python3 setup.py sdist # create package
+    twine upload ${DIST}/* # upload to testpypi
+else
+    echo "Installing new package... "
+    $PIP uninstall ${PACKAGE}
+    $PIP install ${PACKAGE} # install
+    # $PIP uninstall ${PACKAGE}
+fi
